+++ src/snowflake/snowpark/_internal/analyzer/analyzer.py
@@ -121,11 +121,13 @@ from snowflake.snowpark._internal.analyzer.unary_plan_node import (
     Aggregate,
     CreateDynamicTableCommand,
     CreateViewCommand,
+    Exclude,
     Filter,
     LocalTempView,
     PersistedView,
     Pivot,
     Project,
+    Rename,
     Sample,
     Sort,
     Unpivot,
@@ -937,6 +939,20 @@ class Analyzer:
                     self.analyze(c, df_aliased_col_name_to_real_col_name)
                     for c in logical_plan.column_list
                 ],
+                resolved_children[logical_plan.child],
+                logical_plan,
+            )
+
+        if isinstance(logical_plan, Exclude):
+            return self.plan_builder.exclude(
+                logical_plan.column_list,
+                resolved_children[logical_plan.child],
+                logical_plan,
+            )
+
+        if isinstance(logical_plan, Rename):
+            return self.plan_builder.rename(
+                logical_plan.column_map,
                 resolved_children[logical_plan.child],
                 logical_plan,
             )