+++ CHANGELOG.md
@@ -25,7 +25,8 @@
 - Add the alias `Column.contains` for `functions.contains`.
 - Add support for `StructType.add` to append more fields to existing `StructType` objects.
 - Added support for aliases for commonly used options in DataFrame reader/writer.
-- Added support for `format(...).load()` and `format(...).save(...)`patterns.
+- Added support for `format(...).load()` pattern in dataframe reader and `format(...).save(...)` pattern in dataframe writer.
+- Added support for `df.write.csv("path/to/stage")`,`df.write.json("path/to/stage")`,`df.write.parquet("path/to/stage")` patterns to allow a similar experience to dataframe readers.
 
 ### Bug Fixes