+++ CHANGELOG.md
@@ -12,6 +12,10 @@
 
 ## 1.5.0 (TBD)
 
+### Behavior Changes
+
+- Aggregation results, from functions such as `DataFrame.agg` and `DataFrame.describe`, no longer strip away non-printing characters from column names.
+
 ### New Features
 
 - Added support for new functions in `snowflake.snowpark.functions`:
@@ -24,6 +28,7 @@
 - Added the alias `functions.element_at()` for `functions.get()`.
 - Set default reviewer.
 - Add the alias `Column.contains` for `functions.contains`.
+- Add support for `StructType.add` to append more fields to existing `StructType` objects.
 
 ### Bug Fixes