+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.5.0 (TBD)
 
+### Behavior Changes
+
+- Aggregation results, from functions such as `DataFrame.agg` and `DataFrame.describe`, no longer strip away non-printing characters from column names.
+
 ### New Features
 
 - Added support for new functions in `snowflake.snowpark.functions`:
@@ -15,6 +19,7 @@
 - Set default reviewer.
 - Add the alias `Column.contains` for `functions.contains`.
 - Added support for querying metadata columns from stage when creating `DataFrame` using `DataFrameReader`.
+- Add support for `StructType.add` to append more fields to existing `StructType` objects.
 
 ### Bug Fixes