+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Added the alias `functions.element_at()` for `functions.get()`.
 - Set default reviewer.
 - Add the alias `Column.contains` for `functions.contains`.
+- Added support for querying metadata columns from stage when creating `DataFrame` using `DataFrameReader`.
 - Add support for `StructType.add` to append more fields to existing `StructType` objects.
 
 ### Bug Fixes