+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Added support for aliases for commonly used options in DataFrame reader/writer.
 - Added support for `format(...).load()` pattern in dataframe reader and `format(...).save(...)` pattern in dataframe writer.
 - Added support for `df.write.csv("path/to/stage")`,`df.write.json("path/to/stage")`,`df.write.parquet("path/to/stage")` patterns to allow a similar experience to dataframe readers.
+- Added the alias `functions.element_at()` for `functions.get()`
 
 ### Bug Fixes